% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{clear_log_file}
\alias{clear_log_file}
\title{Clear log output file}
\usage{
clear_log_file()
}
\description{
clears the logging file
}
\examples{
clearfile()
}
\keyword{debug}
\keyword{log}
\keyword{print}

