% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_numeric.R
\name{convert_to_numeric}
\alias{convert_to_numeric}
\title{Convert To Numeric}
\usage{
convert_to_numeric(d)
}
\arguments{
\item{d}{the data to be converted}
}
\description{
This function converts a vector or data frame to numeric values
}
\examples{
convert_to_numeric(data.frame(y=array("1", "2", "3")))
}
\keyword{convert}
\keyword{numeric}

