% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_string.R
\name{convert_to_string}
\alias{convert_to_string}
\title{Convert To String}
\usage{
convert_to_string(d)
}
\arguments{
\item{d}{the data to be converted}
}
\description{
This function converts a vector or data frame to string values
}
\examples{
convert_to_string(data.frame(y=rbind(1, 2, 3)))
}
\keyword{convert}
\keyword{string}

