% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jnn.R
\name{jnn}
\alias{jnn}
\title{Jeneral Neural Net}
\usage{
jnn(X, Y, ...)
}
\arguments{
\item{X}{the input matrix}

\item{Y}{the output input matrix}
}
\description{
This function is a custom implementation of a neural network
}
\examples{
jnn(X, Y)
}
\keyword{jeneral}
\keyword{net}
\keyword{network}
\keyword{neural}

