% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurtosis.R
\name{kurtosis}
\alias{kurtosis}
\title{Find kurtosis of a distribution}
\usage{
kurtosis(x)
}
\arguments{
\item{x}{the data to find the kurtosis of}
}
\description{
This function takes in a set of values from a gaussian distribution and returns the kurtosis of that distribution
}
\examples{
kurtosis(x)
}
\keyword{kurtosis}

