% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_predictors.R
\name{merge_predictors}
\alias{merge_predictors}
\title{Merge Predictors}
\usage{
merge_predictors(y, pred1, pred2)
}
\arguments{
\item{y}{the target matrix}

\item{pred1}{the matrix of predictions given by the first model}

\item{pred2}{the matrix of predictions given by the second model}
}
\description{
This function merges multiple predictors to find the optimal combination for prediction
}
\examples{
merge(y, predict(model1, newdata=x), predict(model2, newdata=x))
}
\keyword{merge}
\keyword{predict}
\keyword{predictor}

