% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_log_loss.R
\name{multi_log_loss}
\alias{multi_log_loss}
\title{MultiLogLoss}
\usage{
multi_log_loss(act, pred)
}
\arguments{
\item{act}{the actual observed values for the data set}

\item{pred}{the predicted probabilities for the data set}
}
\description{
This function returns the logloss cost function value for actual and predicted values
}
\examples{
multi_log_loss(y, predict(model, newdata=x))
}
\keyword{log}
\keyword{loss}
\keyword{multilogloss}

