% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naref.R
\name{naref}
\alias{naref}
\title{naref}
\usage{
naref(DF)
}
\arguments{
\item{DF}{the data frame to set with the default NA reference level}
}
\description{
make missing (NA) the reference level of a factor
}
\examples{
naref(DF)
}
\keyword{factor}
\keyword{na}
\keyword{naref}
\keyword{reference}

