% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_predict.R
\name{predict_a}
\alias{predict_a}
\title{Predict a value based on probability and cost of misprediction}
\usage{
predict_a(c, p)
}
\arguments{
\item{c}{the costs matrix}

\item{p}{the probability matrix}
}
\description{
This function takes in a matrix of costs for mispredicting and a matrix of probabilities of various outcomes.  It returns TRUE if you should predict a, and FALSE if you should predict not-a
}
\examples{
predict_a(c, p)
}
\keyword{cost}
\keyword{predict}
\keyword{probability}

