% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_formula.R
\name{print_formula_from_coefs}
\alias{print_formula_from_coefs}
\title{Print Formula From Coefficients}
\usage{
print_formula_from_coefs(coefs)
}
\arguments{
\item{coefs}{the coefficients to be printed}
}
\description{
This function prints the fully expanded formula from the coeficients
}
\examples{
print_formula_from_coefs(reg$coefs)
}
\keyword{coef}
\keyword{coefficients}
\keyword{formula}
\keyword{print}

