% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{printf}
\alias{printf}
\title{Print logging information}
\usage{
printf(content, file = TRUE, debug = FALSE, console = TRUE, ...)
}
\arguments{
\item{content}{the content to print}

\item{file}{whether to print to a file}

\item{debug}{whether this is debug output}

\item{console}{whether to print to console}
}
\description{
Prints content to the console, debug, and file outputs
}
\examples{
printf("hello", TRUE, FALSE, TRUE)
}
\keyword{debug}
\keyword{log}
\keyword{print}

