% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_naref.R
\name{reverse_naref}
\alias{reverse_naref}
\title{reverse naref}
\usage{
reverse_naref(DF)
}
\arguments{
\item{DF}{the data frame to set NA as string}
}
\description{
replace (NA) with "NA"
}
\examples{
reverse_naref(DF)
}
\keyword{factor}
\keyword{na}
\keyword{naref}
\keyword{reference}
\keyword{reverse}

