% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{roc}
\alias{roc}
\title{Plot the ROC curve}
\usage{
roc(p, y, ...)
}
\arguments{
\item{p}{the probability vector}

\item{y}{the response vector}
}
\description{
plot the ROC curve for classification of y with p
}
\examples{
roc(p, y, ...)
}
\keyword{roc}
\keyword{sensitivity}
\keyword{specificity}

