% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{set_debug_print}
\alias{set_debug_print}
\title{Set global debug value}
\usage{
set_debug_print(x)
}
\arguments{
\item{x}{TRUE or FALSE, whether or not to print debug messages}
}
\description{
sets debugging to on for logging purposes
}
\examples{
set_debug_print(TRUE)
}
\keyword{debug}
\keyword{log}
\keyword{print}

