% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{set_log_file}
\alias{set_log_file}
\title{Set log output file}
\usage{
set_log_file(file)
}
\arguments{
\item{file}{the local file path to print to in the working directory}
}
\description{
sets debugging to on for logging purposes
}
\examples{
set_log_file(file)
}
\keyword{debug}
\keyword{log}
\keyword{print}

