% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{set_up_log}
\alias{set_up_log}
\title{Set up log to allow us to use printf()}
\usage{
set_up_log(file_name, debug)
}
\arguments{
\item{file_name}{the name of the file to print to}

\item{debug}{whether to print debut output}
}
\description{
Prints content to the console, debug, and file outputs
}
\examples{
set_up_log("log.txt", TRUE)
}
\keyword{debug}
\keyword{log}
\keyword{print}

