% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xg_fit.R
\name{xg_fit}
\alias{xg_fit}
\title{XG Fit}
\usage{
xg_fit(max_depth_vals = c(5), min_child_weight_vals = c(1), nrounds = 200,
  subsample_vals = c(0.6), eta_vals = c(0.3),
  colsample_bytree_vals = c(0.2), ...)
}
\arguments{
\item{max_depth_vals}{the values for max_depth to take}

\item{min_child_weight_vals}{the values for min_child_weight to take}

\item{nrounds}{the number of trees to build and cross validate}

\item{subsample_vals}{the values for subsample to take}

\item{eta_vals}{the values for eta to take}

\item{colsample_bytree_vals}{the values for colsample_bytree to take}

\item{y}{the target matrix}
}
\description{
This function iterates over the selected parameter space to find the minimal OOS loss using xgboost cross validation
}
\examples{
xgfit()
}
\keyword{fit}
\keyword{xgboost}

